Refactoring Truthtable implementation out of Condition - passing all test cases
