+++ lib/package/FlowPhase.js
@@ -11,7 +11,8 @@ function FlowPhase(element, parent) {
 
 FlowPhase.prototype.getType = function() {
   if (!this.type) {
-    this.type = this.element.parentNode.tagName;
+    this.type="FlowPhase";
+    //this.type = this.getPhase();
   }
   return this.type;
 };
