+++ lib/package/Policy.js
@@ -168,17 +168,19 @@ Policy.prototype.getSteps = function() {
     bundle.getEndpoints().forEach(function(ep) {
       //flows
       ep.getAllFlows().forEach(function(fl) {
-        var fps = [fl.getFlowRequest()];
-        fps.push(fl.getFlowResponse());
-        fps.forEach(function(fp) {
-          if (fp) {
-            fp.getSteps().forEach(function(st) {
-              if (st.getName() === policyName) {
-                steps.push(st);
-              }
-            });
-          }
-        });
+        if (fl) {
+          var fps = [fl.getFlowRequest()];
+          fps.push(fl.getFlowResponse());
+          fps.forEach(function(fp) {
+            if (fp) {
+              fp.getSteps().forEach(function(st) {
+                if (st.getName() === policyName) {
+                  steps.push(st);
+                }
+              });
+            }
+          });
+        }
       });
       //faultrules
       ep.getFaultRules().forEach(function(fr) {
