+++ lib/package/plugins/checkConditionComplexity.js
@@ -11,7 +11,7 @@ var plugin = {
   threshold = 12;
 
 var onCondition = function(condition) {
-  var ast = condition.getAST(),
+  var ast = condition.getTruthTable().getAST(),
     nodeCount = 0,
     nodes = [ast];
