+++ lib/package/plugins/checkConditionForLiterals.js
@@ -12,10 +12,10 @@ var plugin = {
 };
 
 var onCondition = function(condition) {
-  var ast = condition.getAST(),
+  var ast = condition.getTruthTable().getAST(),
     hasLiteral = false,
     nodes = [ast],
-    actions = ["root"];
+    actions = [{ action: "root", parent: undefined }];
 
   while (nodes[0] && !hasLiteral) {
     var node = nodes.pop(),
@@ -23,10 +23,11 @@ var onCondition = function(condition) {
 
     if (
       node.type === "constant" &&
-      (!parentAction ||
-        parentAction.match(
-          /^(root|conjunction|disjunction|negation|implication)$/
-        ))
+      (!parentAction.parent ||
+        (parentAction.parent &&
+          parentAction.parent.match(
+            /^(root|conjunction|disjunction|negation|implication)$/
+          )))
     ) {
       hasLiteral = true;
       //write error
@@ -36,7 +37,13 @@ var onCondition = function(condition) {
         message: "Condition contains literal outside of comparator."
       });
     } else if (node.args) {
-      actions.push(node.action);
+      //lets also capture the prior action
+      var act = { action: node.action, parent: undefined };
+      if (parentAction && parentAction.action) {
+        act.parent = parentAction.action;
+      }
+
+      actions.push(act);
       Array.prototype.push.apply(nodes, node.args);
     }
   }
