+++ lib/package/plugins/checkConditionTruth.js
@@ -7,7 +7,7 @@ var plugin = {
     fatal: false,
     severity: 1,
     nodeType:"Condition",
-    enabled:false
+    enabled:true
   },
   myUtil = require("../myUtil.js");
