+++ lib/package/plugins/extractJSONConditionCheck.js
@@ -34,8 +34,12 @@ var onPolicy = function(policy) {
         if (!condition || !condition.getExpression().match(condRegExp)) {
           missingBodyCheck = true;
 
+          var parent=step.parent;
           //is the parent a flow we might revert the decision if it has an appropriate condition
-          if (step.parent && step.parent.getType() === "Flow") {
+          if(parent.getType()==="FlowPhase"){
+            parent=parent.parent;
+          }
+          if (parent && parent.getType() === "Flow" ) {//&& step.parent.getCondition
             var condition = step.parent.getCondition();
             if (condition && condition.getExpression().match(condRegExp)) {
               missingBodyCheck = false;
