+++ test/testcheckConditionTruth.js
